+++ sdk/src/adminClient.ts
@@ -3960,14 +3960,56 @@ export class AdminClient extends DriftClient {
 		);
 	}
 
+	public async updatePerpMarketOracleSlotDelayOverride(
+		perpMarketIndex: number,
+		oracleSlotDelay: number
+	): Promise<TransactionSignature> {
+		const updatePerpMarketOracleSlotDelayOverrideIx =
+			await this.getUpdatePerpMarketOracleSlotDelayOverrideIx(
+				perpMarketIndex,
+				oracleSlotDelay
+			);
+		const tx = await this.buildTransaction(
+			updatePerpMarketOracleSlotDelayOverrideIx
+		);
+		const { txSig } = await this.sendTransaction(tx, [], this.opts);
+
+		return txSig;
+	}
+
+	public async getUpdatePerpMarketOracleSlotDelayOverrideIx(
+		perpMarketIndex: number,
+		oracleSlotDelay: number
+	): Promise<TransactionInstruction> {
+		const perpMarketPublicKey = await getPerpMarketPublicKey(
+			this.program.programId,
+			perpMarketIndex
+		);
+
+		return await this.program.instruction.updatePerpMarketOracleSlotDelayOverride(
+			oracleSlotDelay,
+			{
+				accounts: {
+					admin: this.useHotWalletAdmin
+						? this.wallet.publicKey
+						: this.getStateAccount().admin,
+					state: await this.getStatePublicKey(),
+					perpMarket: perpMarketPublicKey,
+				},
+			}
+		);
+	}
+
 	public async updatePerpMarketAmmSpreadAdjustment(
 		perpMarketIndex: number,
-		ammSpreadAdjustment: number
+		ammSpreadAdjustment: number,
+		ammInventorySpreadAdjustment: number
 	): Promise<TransactionSignature> {
 		const updatePerpMarketAmmSpreadAdjustmentIx =
 			await this.getUpdatePerpMarketAmmSpreadAdjustmentIx(
 				perpMarketIndex,
-				ammSpreadAdjustment
+				ammSpreadAdjustment,
+				ammInventorySpreadAdjustment
 			);
 		const tx = await this.buildTransaction(
 			updatePerpMarketAmmSpreadAdjustmentIx
@@ -3979,7 +4021,8 @@ export class AdminClient extends DriftClient {
 
 	public async getUpdatePerpMarketAmmSpreadAdjustmentIx(
 		perpMarketIndex: number,
-		ammSpreadAdjustment: number
+		ammSpreadAdjustment: number,
+		ammInventorySpreadAdjustment: number
 	): Promise<TransactionInstruction> {
 		const perpMarketPublicKey = await getPerpMarketPublicKey(
 			this.program.programId,
@@ -3988,6 +4031,7 @@ export class AdminClient extends DriftClient {
 
 		return await this.program.instruction.updatePerpMarketAmmSpreadAdjustment(
 			ammSpreadAdjustment,
+			ammInventorySpreadAdjustment,
 			{
 				accounts: {
 					admin: this.useHotWalletAdmin